+++ .eslintrc
@@ -68,8 +68,6 @@
     "react/no-array-index-key": "off",
     "react/self-closing-comp": "off",
     "@typescript-eslint/no-explicit-any": "error",
-    "jsx-a11y/anchor-is-valid": "off",
-    "jsx-a11y/href-no-hash": "off",
     "default-case": "off",
     "max-lines": [
       "error",
