+++ src/routes/Monsters/SampleModal.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable jsx-a11y/anchor-is-valid */
 import React from 'react'
 import styled from '@emotion/styled'
 import { Modal, Button } from 'antd'
