+++ src/store/monsters/actions.ts
@@ -1,5 +1,5 @@
 import { Dispatch } from '~/store/types'
-import { ActionTypes, HOME_ACTIONS } from './types'
+import { ActionTypes, Monster, HOME_ACTIONS } from './types'
 import { get, ENDPOINTS } from '~/store/api'
 
 export const getMonsters = () => (dispatch: Dispatch<ActionTypes>) => {
@@ -8,12 +8,14 @@ export const getMonsters = () => (dispatch: Dispatch<ActionTypes>) => {
     isLoading: true,
   })
 
-  get(ENDPOINTS.MONSTERS_ENDPOINT).then((resp) => {
-    dispatch({
-      type: HOME_ACTIONS.GET_MONSTERS,
-      monsters: resp.results,
+  get(ENDPOINTS.MONSTERS_ENDPOINT)
+    .then((resp: { results: Monster[] }) => {
+      dispatch({
+        type: HOME_ACTIONS.GET_MONSTERS,
+        monsters: resp.results,
+      })
     })
-  })
+    .catch(() => {})
 }
 
 export const setCurrentPage = (currentPage: number) => (dispatch: Dispatch<ActionTypes>) => {
