+++ src/store/monsters/types.ts
